requirements.txt
aws-cdk-lib==2.17.0
constructs>=10.0.0,<11.0.0
boto3=1.21.25
cdk-dynamo-table-view=0.2.105